refactor: revert EVM gas type to uint64_t for Wasm alignment

+++ src/evm/interpreter.cpp
@@ -64,7 +64,8 @@ void InterpreterExecContext::freeBackFrame() {
   auto &BackFrame = FrameStack.back();
 
   GasUsed = GasUsed - BackFrame.Msg->gas;
-  int64_t GasRefund = std::min(BackFrame.GasRefund, BackFrame.Msg->gas / 2);
+  uint64_t GasRefund = std::min(
+      BackFrame.GasRefund, static_cast<uint64_t>(BackFrame.Msg->gas / 2LL));
   GasUsed = GasUsed - GasRefund;
 
   FrameStack.pop_back();
